Refactor reversion history cleanup imports to new history_cleanup module for better organization
