+++ boranga/components/data_migration/handlers/species.py
@@ -85,7 +85,7 @@ class SpeciesImporter(BaseSheetImporter):
             report_filter = {}
 
         # Delete reversion history first (more efficient than waiting for cascade)
-        from boranga.components.data_migration.utils.reversion_cleanup import ReversionHistoryCleaner
+        from boranga.components.data_migration.history_cleanup.reversion_cleanup import ReversionHistoryCleaner
 
         cleaner = ReversionHistoryCleaner(batch_size=2000)
         cleaner.clear_species_and_related(species_filter)
